+++ backend/resources/club_routes.py
@@ -1,7 +1,7 @@
 #pylint: disable-all
 
 from flask import Blueprint, Response, request, jsonify
-from database.models import Clubs, ClubSchema, club_schema, db, ma
+from database.models import Clubs, ClubSchema, club_schema, ClubConvertor, db, ma
 
 clubblue = Blueprint("clubblue", __name__)
 
@@ -11,7 +11,7 @@ def add_club():
     SID = request.json['SID']
     Club_codes = request.json['Club_codes']
 
-    new_club = Clubs(SID, Club_codes, SID)
+    new_club = Clubs(SID, Club_codes)
 
     db.session.add(new_club)
     db.session.commit()
@@ -24,7 +24,11 @@ def add_club():
 @clubblue.route('/signupclub', methods=["POST"])
 def clubs_add():
     SID = request.json['SID']
-    club_list = request.json['Clubs']
+    club_list = request.json['Clubs'] # as a list
+
+    if club_list is None :
+        # empty list
+        return jsonify({'code':200})
 
     club_string = ""
     for club in club_list:
@@ -46,6 +50,10 @@ def update_club(SID):
     SID = request.json['SID']
     Club_codes = request.json['Club_codes']
 
+    if not club_detail :
+        # student has no current clubs, make a POST request
+        return jsonify({'code':501})
+
     if len(Club_codes) == 0 :
         db.session.delete(club_detail)
         db.session.commit()
@@ -61,4 +69,22 @@ def update_club(SID):
 
 
 # GET a list of clubs for a particular SID
+@clubblue.route('/club/<SID>', methods =['GET'])
+def get_clubs(SID):
+    club_detail = Clubs.query.get(SID)
 
+    clubs_of_student = club_detail.Club_codes
+
+    if not club_detail :
+        # student has no club
+        return jsonify({'code': 407})
+
+    club_list = []
+    for i in range(0,len(clubs_of_student),2):
+        club_str = clubs_of_student[i:i+2]
+
+        #convert names into codes
+        club_name = ClubConvertor.query.get(club_str).Club
+        club_list.append(club_name)
+
+    return jsonify({'clubs' : club_list})
